remove additional default/theme actions
added a function to handle removal of actions across various potential
hooks and created a simple array to run the removal function against
several possible actions that may interfere with default output.

The invocation is contained to the template files so this is easily
altered via child themes using custom template files.